+++ projects/np-ui-lib/src/lib/np-list/np-list.module.ts
@@ -3,13 +3,17 @@ import { CommonModule } from '@angular/common';
 import { NpListComponent } from './np-list.component';
 import { NpUtilityModule } from '../np-utility/np-utility.module';
 import { NpPaginatorModule } from '../np-paginator/np-paginator.module';
+import { NpCheckboxModule } from '../np-checkbox/np-checkbox.module';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [NpListComponent],
   imports: [
     CommonModule,
+    FormsModule,
     NpUtilityModule,
-    NpPaginatorModule
+    NpPaginatorModule,
+    NpCheckboxModule
   ],
   exports: [NpListComponent]
 })