+++ projects/np-ui-lib/src/lib/np-tree-view/np-tree-view.module.ts
@@ -1,13 +1,15 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { NpTreeViewComponent } from './np-tree-view.component';
-import { NpUtilityModule } from '../np-utility/np-utility.module';
+import { NpCheckboxModule } from '../np-checkbox/np-checkbox.module';
+import { FormsModule } from '@angular/forms';
 
 @NgModule({
   declarations: [NpTreeViewComponent],
   imports: [
     CommonModule,
-    NpUtilityModule
+    FormsModule,
+    NpCheckboxModule
   ],
   exports: [NpTreeViewComponent]
 })