+++ projects/np-ui-lib/src/public-api.ts
@@ -42,7 +42,6 @@ export * from './lib/np-panel/np-panel-content.directive';
 export * from './lib/np-utility/np-utility.module';
 export * from './lib/np-utility/np-highlight.pipe';
 export * from './lib/np-utility/np-mask.directive';
-export * from './lib/np-utility/np-indeterminate.directive';
 export * from './lib/np-utility/np-orderby.pipe';
 export * from './lib/np-utility/np-autofocus.directive';